# api/swagger_ui/api.yml
openapi: 3.0.0
info:
  title: iothub
  version: 1.0.0
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
paths:
  /api/v1/things/{thingId}/methods/switch:
    post:
      tags:
        - default
      summary: Invoke Method
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                respTimeout: 30
                data:
                  - chnNo: 1
                    onoff: 1
      security:
        - basicAuth: []
      parameters:
        - name: thingId
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          headers:
            Date:
              schema:
                type: string
                example: Wed, 14 Sep 2022 13:31:51 GMT
            Content-Length:
              schema:
                type: integer
                example: '79'
            Content-Type:
              schema:
                type: string
                example: text/plain; charset=utf-8
          content:
            application/json:
              schema:
                type: object
              examples:
                example-0:
                  summary: '200'
                  value:
                    code: 200
                    message: OK
                    data:
                      status: 200
                      message: OK
                      data:
                        a: 1
                example-1:
                  summary: '504'
                  value:
                    code: 504
                    message: 'wait 30 seconds: method timeout'
                    data: null
  /api/v1/things:
    post:
      tags:
        - default
      summary: Add Thing
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                thingId: abc
      security:
        - basicAuth: []
      responses:
        '200':
          description: OK
          headers:
            Connection:
              schema:
                type: string
                example: close
            Content-Length:
              schema:
                type: integer
                example: '230'
            Content-Type:
              schema:
                type: string
                example: application/json
            Date:
              schema:
                type: string
                example: Mon, 26 Sep 2022 09:32:42 GMT
          content:
            application/json:
              schema:
                type: object
              example:
                code: 200
                message: OK
                data:
                  thingId: test3
                  enabled: true
                  authType: password
                  authValue: 72ab2f58-6303-437d-ae98-cc3c33fe34d9
                  updateAt: '2022-09-26T09:32:42.667173616Z'
                  createAt: '2022-09-26T09:32:42.667173616Z'
        '400':
          description: Bad Request
          headers:
            Date:
              schema:
                type: string
                example: Wed, 14 Sep 2022 13:13:38 GMT
            Content-Length:
              schema:
                type: integer
                example: '56'
            Content-Type:
              schema:
                type: string
                example: text/plain; charset=utf-8
          content:
            application/json:
              schema:
                type: object
              example:
                code: 400
                message: entity already exists
                data: ''
  /api/v1/things/{thingId}:
    get:
      tags:
        - default
      summary: Get Thing
      security:
        - basicAuth: []
      parameters:
        - name: thingId
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          headers:
            Date:
              schema:
                type: string
                example: Wed, 14 Sep 2022 14:04:08 GMT
            Content-Length:
              schema:
                type: integer
                example: '234'
            Content-Type:
              schema:
                type: string
                example: text/plain; charset=utf-8
          content:
            application/json:
              schema:
                type: object
              example:
                code: 200
                message: OK
                data:
                  thingId: test1
                  enabled: true
                  authType: password
                  authValue: 79bc1216-e605-4b96-b4e3-ad0f62350e61
                  updateAt: '2022-09-14T21:13:35.804559+08:00'
                  createAt: '2022-09-14T21:13:35.804559+08:00'
    delete:
      tags:
        - default
      summary: Delete Thing
      security:
        - basicAuth: []
      parameters:
        - name: thingId
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
            Date:
              schema:
                type: string
                example: Fri, 23 Sep 2022 07:17:24 GMT
            Content-Length:
              schema:
                type: integer
                example: '37'
          content:
            application/json:
              schema:
                type: object
              example:
                code: 200
                message: OK
                data: ''
  /api/v1/things/{thingId}/shadows/default:
    get:
      tags:
        - default
      summary: Get Shadow
      security:
        - basicAuth: []
      parameters:
        - name: thingId
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          headers:
            Connection:
              schema:
                type: string
                example: close
            Content-Length:
              schema:
                type: integer
                example: '557'
            Content-Type:
              schema:
                type: string
                example: application/json
            Date:
              schema:
                type: string
                example: Mon, 26 Sep 2022 10:03:40 GMT
          content:
            application/json:
              schema:
                type: object
              example:
                code: 200
                message: OK
                data:
                  connected: true
                  thingId: test1
                  state:
                    desired:
                      config:
                        sunrise: '17:43'
                        sunset: '5:44'
                    reported:
                      chnNo_1:
                        bri: 53
                        onoff: 1
                        srcInfo: null
                        srcType: CAT1
                      chnNo_2:
                        bri: 53
                        onoff: 0
                        srcInfo: null
                        srcType: CAT1
                      reportPeriodOfSec: 300
                  metadata:
                    desired:
                      config:
                        timestamp: 1664125202635
                    reported:
                      chnNo_1:
                        timestamp: 1663924107827
                      chnNo_2:
                        timestamp: 1663924107827
                  version: 6122
                  updatedAt: '2022-09-26T09:07:49.884435175Z'
                  createdAt: '2022-09-14T14:46:52.142508333Z'
  /api/v1/things/{thingId}/shadows/default/state/desired:
    put:
      tags:
        - default
      summary: Set Shadow Desired
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                clientToken: sssdsd
                state:
                  desired:
                    xxx: 1
      security:
        - basicAuth: []
      parameters:
        - name: thingId
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
            Date:
              schema:
                type: string
                example: Fri, 23 Sep 2022 07:11:51 GMT
            Content-Length:
              schema:
                type: integer
                example: '37'
          content:
            application/json:
              schema:
                type: object
              example:
                code: 200
                message: OK
                data: ''